i18nのindex.jsをdynamicとstaticとmdxに分けた
