+++ frontend/app/[locale]/layout.tsx
@@ -3,7 +3,7 @@ import "@fontsource/kaisei-opti/japanese-400.css";
 import "@fontsource/noto-sans/400.css";
 import "@fontsource/noto-sans-jp/japanese-400.css";
 import "@/globals.css";
-import { getMessages, locales } from "@falling-nikochan/i18n";
+import { getMessages, locales } from "@falling-nikochan/i18n/dynamic";
 import IntlProvider from "./intlProvider.js";
 import { initMetadata, initViewport, MetadataProps } from "./metadata.js";
 import { ThemeProvider } from "./common/theme.jsx";
