+++ frontend/app/[locale]/main/about/[aboutIndex]/1-about.tsx
@@ -1,5 +1,5 @@
 import { PWAInstallDesc } from "@/common/pwaInstall";
-import { getTranslations } from "@falling-nikochan/i18n";
+import { getTranslations } from "@falling-nikochan/i18n/dynamic";
 
 export async function AboutContent1({locale}: {locale:string}) {
   const t = await getTranslations(locale, "about.1");
