+++ frontend/app/[locale]/main/about/[aboutIndex]/page.tsx
@@ -5,7 +5,7 @@ import { AboutContent3 } from "./3-edit.js";
 import { Pager } from "@/common/pager.js";
 import { AboutContent4 } from "./4-level.js";
 import { AboutContent5 } from "./5-judge.js";
-import { getTranslations } from "@falling-nikochan/i18n";
+import { getTranslations } from "@falling-nikochan/i18n/dynamic";
 import { initMetadata } from "@/metadata.js";
 import { maxAboutPageIndex } from "./pager.js";
