+++ frontend/app/[locale]/main/links/page.tsx
@@ -1,6 +1,6 @@
 import LinksPage from "./clientPage.js";
 import { initMetadata, MetadataProps } from "@/metadata.js";
-import { getTranslations } from "@falling-nikochan/i18n";
+import { getTranslations } from "@falling-nikochan/i18n/dynamic";
 
 export async function generateMetadata({ params }: MetadataProps) {
   const t = await getTranslations(params, "main.links");
