+++ route/src/share.ts
@@ -1,5 +1,5 @@
 import { Bindings, cacheControl, languageDetector } from "./env.js";
-import { getTranslations } from "@falling-nikochan/i18n";
+import { getTranslations } from "@falling-nikochan/i18n/dynamic.js";
 import {
   baseScoreRate,
   bigScoreRate,
